+++ src/ivoc/ocjump.cpp
@@ -1,5 +1,6 @@
 #include <../../nrnconf.h>
 
+#include "nrnfilewrap.h"
 #include "nrnoc2iv.h"
 #include "ocfunc.h"
 #include "ocjump.h"
@@ -20,7 +21,7 @@ extern int hoc_execerror_messages;
 bool hoc_valid_stmt(const char* stmt, Object* ob) {
     std::string s{stmt};
     s.append(1, '\n');
-    return OcJump{}.execute(s.c_str(), ob);
+    return OcJump::execute(s.c_str(), ob);
 }
 
 void hoc_execute1() {
@@ -47,7 +48,7 @@ void hoc_execute1() {
 
 #if HAVE_IV
 bool Oc::valid_expr(Symbol* s) {
-    return OcJump{}.execute(s->u.u_proc->defn.in);
+    return OcJump::execute(s->u.u_proc->defn.in);
 }
 
 bool Oc::valid_stmt(const char* stmt, Object* ob) {
@@ -57,63 +58,88 @@ bool Oc::valid_stmt(const char* stmt, Object* ob) {
 //------------------------------------------------------------------
 void hoc_execute(Inst*);
 
+namespace {
+struct saved_state {
+    saved_state() {
+        // not complete but it is good for expressions and it can be improved
+        oc_save_hoc_oop(&o1, &o2, &o4, &o5);
+        oc_save_code(&c1, &c2, &c3, &c4, &c5, &c6, &c7, &c8, &c9, &c10, &c11, &c12);
+        oc_save_input_info(&i1, &i2, &i3, &i4);
+        oc_save_cabcode(&cc1, &cc2);
+    }
+    void restore() {
+        oc_restore_hoc_oop(&o1, &o2, &o4, &o5);
+        oc_restore_code(&c1, &c2, &c3, &c4, &c5, &c6, &c7, &c8, &c9, &c10, &c11, &c12);
+        oc_restore_input_info(i1, i2, i3, i4);
+        oc_restore_cabcode(&cc1, &cc2);
+    }
+
+  private:
+    // hoc_oop
+    Object* o1{};
+    Objectdata* o2{};
+    int o4{};
+    Symlist* o5{};
+
+    // code
+    Inst* c1{};
+    Inst* c2{};
+    Datum* c3{};
+    nrn::oc::frame* c4{};
+    int c5{};
+    int c6{};
+    Inst* c7{};
+    nrn::oc::frame* c8{};
+    Datum* c9{};
+    Symlist* c10{};
+    Inst* c11{};
+    int c12{};
+
+    // input_info
+    const char* i1{};
+    int i2{};
+    int i3{};
+    NrnFILEWrap* i4{};
+
+    // cabcode
+    int cc1{};
+    int cc2{};
+};
+}  // namespace
+
 bool OcJump::execute(Inst* p) {
-    bool ret{false};
-    begin();
+    saved_state before{};
+    try_catch_depth_increment tell_children_we_will_catch{};
     try {
         hoc_execute(p);
-        ret = true;
+        return true;
     } catch (...) {
-        restore();
+        before.restore();
+        return false;
     }
-    finish();
-    return ret;
 }
 
 bool OcJump::execute(const char* stmt, Object* ob) {
-    bool ret{false};
-    begin();
+    saved_state before{};
+    try_catch_depth_increment tell_children_we_will_catch{};
     try {
         hoc_obj_run(stmt, ob);
-        ret = true;
+        return true;
     } catch (...) {
-        restore();
+        before.restore();
+        return false;
     }
-    finish();
-    return ret;
 }
 
 void* OcJump::fpycall(void* (*f)(void*, void*), void* a, void* b) {
-    void* c = 0;
-    begin();
+    saved_state before{};
+    try_catch_depth_increment tell_children_we_will_catch{};
     try {
-        c = (*f)(a, b);
+        return (*f)(a, b);
     } catch (...) {
-        restore();
+        before.restore();
+        return nullptr;
     }
-    finish();
-    return c;
-}
-
-extern int nrn_try_catch_nest_depth;
-void OcJump::begin() {
-    // not complete but it is good for expressions and it can be improved
-    oc_save_hoc_oop(&o1, &o2, &o4, &o5);
-    oc_save_code(&c1, &c2, &c3, &c4, &c5, &c6, &c7, &c8, &c9, &c10, &c11, &c12);
-    oc_save_input_info(&i1, &i2, &i3, &i4);
-    oc_save_cabcode(&cc1, &cc2);
-    ++nrn_try_catch_nest_depth;
-}
-
-void OcJump::restore() {
-    oc_restore_hoc_oop(&o1, &o2, &o4, &o5);
-    oc_restore_code(&c1, &c2, &c3, &c4, &c5, &c6, &c7, &c8, &c9, &c10, &c11, &c12);
-    oc_restore_input_info(i1, i2, i3, i4);
-    oc_restore_cabcode(&cc1, &cc2);
-}
-
-void OcJump::finish() {
-    --nrn_try_catch_nest_depth;
 }
 
 ObjectContext::ObjectContext(Object* obj) {
