+++ src/ivoc/ocjump.h
@@ -1,14 +1,29 @@
 #pragma once
-#include "nrnfilewrap.h"
-union Datum;
 union Inst;
-struct OcJumpImpl;
 struct Object;
 union Objectdata;
 struct Symlist;
-namespace nrn::oc {
-struct frame;
-}
+
+/** @brief How many NEURON try { ... } catch(...) { ... } blocks are in the call stack.
+ *
+ *  Errors inside NEURON are triggered using hoc_execerror, which ultimately
+ *  throws an exception. To replicate the old logic, we sometimes need to insert
+ *  a try/catch block *only if* there is no try/catch block less deeply nested
+ *  on the call stack. This global variable tracks how many such blocks are
+ *  currently present on the stack.
+ */
+extern int nrn_try_catch_nest_depth;
+
+/** @brief Helper type for incrementing/decrementing nrn_try_catch_nest_depth.
+ */
+struct try_catch_depth_increment {
+    try_catch_depth_increment() {
+        ++nrn_try_catch_nest_depth;
+    }
+    ~try_catch_depth_increment() {
+        --nrn_try_catch_nest_depth;
+    }
+};
 
 struct ObjectContext {
     ObjectContext(Object*);
@@ -22,42 +37,7 @@ struct ObjectContext {
 };
 
 struct OcJump {
-    bool execute(Inst* p);
-    bool execute(const char*, Object* ob = NULL);
-    void* fpycall(void* (*) (void*, void*), void*, void*);
-
-  private:
-    void begin();
-    void restore();
-    void finish();
-
-    // hoc_oop
-    Object* o1{};
-    Objectdata* o2{};
-    int o4{};
-    Symlist* o5{};
-
-    // code
-    Inst* c1{};
-    Inst* c2{};
-    Datum* c3{};
-    nrn::oc::frame* c4{};
-    int c5{};
-    int c6{};
-    Inst* c7{};
-    nrn::oc::frame* c8{};
-    Datum* c9{};
-    Symlist* c10{};
-    Inst* c11{};
-    int c12{};
-
-    // input_info
-    const char* i1{};
-    int i2{};
-    int i3{};
-    NrnFILEWrap* i4{};
-
-    // cabcode
-    int cc1{};
-    int cc2{};
+    static bool execute(Inst* p);
+    static bool execute(const char*, Object* ob = NULL);
+    static void* fpycall(void* (*) (void*, void*), void*, void*);
 };
